LPS-177454 Remove Feature Flag LPS-165491
