+++ modules/apps/questions/questions-web/src/main/resources/META-INF/resources/js/App.es.js
@@ -33,9 +33,7 @@ export default function App(props) {
 
 	const packageName = props.npmResolvedPackageName;
 
-	const questionsComponent = Liferay.FeatureFlags['LPS-165491']
-		? `${packageName}/js/pages/questions/Questions.new.es`
-		: `${packageName}/js/pages/questions/Questions.es`;
+	const questionsComponent = `${packageName}/js/pages/questions/Questions.es`;
 
 	let path = props.historyRouterBasePath;
